libcamera helpers
